fix(make): arch-lint target, tagged vet/lint, schema-verify by definition
- make arch-lint (the script CI already runs) added and wired into make ci.
- vet/lint run a second pass with ALL_TEST_TAGS (integration,e2e). CI's
  quality gate calls make vet / make lint, so the 42 build-tagged test
  files were never vetted or linted. Fixes the 152 issues that surfaced,
  all in test code: 139 unchecked resp.Body.Close, 6 unused helpers
  removed, 7 unparam (3 unused ctx -> _, 4 documented nolint), plus one
  ineffassign and two unconvert. No test behaviour change; make test-ci
  passes.
- schema-verify checks each produced schema's definition is registered +
  AVAILABLE (get-schema-by-definition, the pod's startup lookup) across
  both registries, instead of name-only get-schema. Verified against floci.
- Runbook, README, .claude/CLAUDE.md updated.

+++ test/integration/relay_test.go
@@ -288,7 +288,7 @@ type flakyPublisher struct {
 
 func (f *flakyPublisher) Publish(ctx context.Context, env events.Envelope[json.RawMessage]) error {
 	n := f.calls.Add(1)
-	if n <= int32(f.failFirstN) {
+	if n <= f.failFirstN {
 		return errors.New("simulated transient SNS failure")
 	}
 	return f.inner.Publish(ctx, env)
@@ -296,7 +296,7 @@ func (f *flakyPublisher) Publish(ctx context.Context, env events.Envelope[json.R
 
 func (f *flakyPublisher) PublishBatch(ctx context.Context, envs []events.Envelope[json.RawMessage]) error {
 	n := f.calls.Add(int32(len(envs)))
-	if n <= int32(f.failFirstN) {
+	if n <= f.failFirstN {
 		return errors.New("simulated transient SNS batch failure")
 	}
 	return f.inner.PublishBatch(ctx, envs)
